// apps/web/src/api-client/model/rideStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */

export type RideStatus = (typeof RideStatus)[keyof typeof RideStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RideStatus = {
    PLANNED: "PLANNED",
    IN_PROGRESS: "IN_PROGRESS",
    COMPLETED: "COMPLETED",
    CANCELLED: "CANCELLED",
} as const;